Move webpack utils to `core-webpack`
