+++ lib/core-webpack/src/importPipeline.ts
