+++ lib/core-webpack/src/index.ts
@@ -2,3 +2,5 @@ export * from './types';
 export * from './load-custom-webpack-config';
 export * from './check-webpack-version';
 export * from './merge-webpack-config';
+export * from './to-importFn';
+export * from './to-require-context';
